Break out units into its own module, add msv sf ag, usmc hmtr
