+++ addons/units/CfgVehicles.hpp
@@ -1,0 +1,10 @@
+class CfgVehicles {
+    // add POTATO USMC
+    #include "subClasses\usmc.hpp"
+
+    // add POTATO MSV
+    #include "subClasses\msv.hpp"
+
+    // add POTATO Airborne
+    #include "subClasses\air.hpp"
+};
