+++ addons/units/subStrings/usmc.hpp
@@ -1,0 +1,54 @@
+QUSMC(rifleman),
+QUSMC(rifleman_02),
+QUSMC(rifleman_03),
+QUSMC(rifleman_04),
+QUSMC(sf_rifleman),
+QUSMC(sniper),
+QUSMC(aar),
+QUSMC(sf_aar),
+QUSMC(mmgag),
+QUSMC(hmgag),
+QUSMC(matag),
+QUSMC(hatag),
+QUSMC(msamag),
+QUSMC(mtrag),
+QUSMC(mtrg),
+QUSMC(hmtrag),
+QUSMC(hmtrg),
+QUSMC(ar),
+QUSMC(sf_ar),
+QUSMC(mmgg),
+QUSMC(hmgg),
+QUSMC(lat),
+QUSMC(sf_lat),
+QUSMC(matg),
+QUSMC(hatg),
+QUSMC(msamg),
+QUSMC(coy),
+QUSMC(xo),
+QUSMC(fac),
+QUSMC(plt),
+QUSMC(sgt),
+QUSMC(sl),
+QUSMC(sf_sl),
+QUSMC(mmgl),
+QUSMC(hmgl),
+QUSMC(matl),
+QUSMC(hatl),
+QUSMC(msaml),
+QUSMC(mtrl),
+QUSMC(hmtrl),
+QUSMC(stl),
+QUSMC(ftl),
+QUSMC(sf_ftl),
+QUSMC(vicl),
+QUSMC(cm),
+QUSMC(plm),
+QUSMC(sm),
+QUSMC(sf_sm),
+QUSMC(eodl),
+QUSMC(eod),
+QUSMC(vicc),
+QUSMC(vicd),
+QUSMC(pilot),
+QUSMC(cc),
